This sorta, kinda starts to make sense
A custom MakeWriter provides custom Writers that don't actually write
output, but instead push to a Vec that is shared with an Arc+Mutex
combination so that its content can be examined after executing a piece
of code that does some tracing/logging.

This should now work as a basis for building a setup where I can
actually write tests for the JSON output behaviour and verify that the
fields from all the spans get printed for each logging event.